separate the forked file into file_utils.py
